LRCI-3892 Remove unnecessary BuildInformation component
